+++ test/utopia/redirection.rb
@@ -32,15 +32,14 @@ describe Utopia::Redirection do
 				Utopia::Response[404, {}, []]
 			end
 		}) do
-			use Utopia::Redirection::Rewrite, {"/" => "/welcome/index"}
-			use Utopia::Redirection::DirectoryIndex
-			use Utopia::Redirection::Errors, {
-				404 => "/error",
-				418 => "/teapot"
-			}
-			use Utopia::Redirection::Moved, "/a", "/b"
-			use Utopia::Redirection::Moved, "/hierarchy/", "/hierarchy", flatten: true
-			use Utopia::Redirection::Moved, "/weird", "/status", status: 333
+			use Utopia::Redirection do |redirects|
+				redirects.rewrite "/" => "/welcome/index"
+				redirects.directory_index
+				redirects.moved "/a", "/b"
+				redirects.moved "/hierarchy/", "/hierarchy", flatten: true
+				redirects.moved "/weird", "/status", status: 333
+			end
+			use Utopia::Redirection::Errors, 404 => "/error", 418 => "/teapot"
 		end
 	end
 	
@@ -82,6 +81,27 @@ describe Utopia::Redirection do
 		
 		expect(last_response.status).to be == 404
 		expect(last_response.read).to be == "File not found :("
+	end
+	
+	it "bypasses request redirections for internal error documents" do
+		application = Utopia::Application.build(Protocol::HTTP::Middleware.for do |request|
+			if request.path_info == "/error"
+				Utopia::Response.text("Internal error document")
+			else
+				Utopia::Response[404, {}, []]
+			end
+		end) do
+			use Utopia::Redirection do |redirects|
+				redirects.rewrite "/error" => "/redirected"
+			end
+			use Utopia::Redirection::Errors, 404 => "/error"
+		end
+		
+		response = application.call(Protocol::HTTP::Request["GET", "/missing"])
+		
+		expect(response.status).to be == 404
+		expect(response.headers["location"]).to be == nil
+		expect(response.read).to be == "Internal error document"
 	end
 	
 	it "closes the response replaced by an error document" do
@@ -141,5 +161,18 @@ describe Utopia::Redirection do
 		
 		expect(last_response.status).to be == 333
 		expect(last_response.headers["location"]).to be == "/status"
+	end
+	
+	it "applies request rules in declaration order" do
+		application = Utopia::Application.build do
+			use Utopia::Redirection do |redirects|
+				redirects.rewrite "/files/" => "/exact"
+				redirects.directory_index "index"
+			end
+		end
+		
+		response = application.call(Protocol::HTTP::Request["GET", "/files/"])
+		
+		expect(response.headers["location"]).to be == "/exact"
 	end
 end